Support large EMC values and update notices

+++ src/test/java/net/exohayvan/dissolver_enhanced/config/ModConfigTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 
+import java.math.BigInteger;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.Charset;
@@ -101,9 +102,9 @@ class ModConfigTest {
                 isDefaults();
 
                 assertThat(VisibilityModifier.EMC_OVERRIDES.items())
-                    .containsEntry("test:case", 400)
-                    .containsEntry("minecraft:dirt", 200)
-                    .containsEntry("minecraft:cobblestone", 300);
+                    .containsEntry("test:case", BigInteger.valueOf(400))
+                    .containsEntry("minecraft:dirt", BigInteger.valueOf(200))
+                    .containsEntry("minecraft:cobblestone", BigInteger.valueOf(300));
 
                 break;
             case INVERTED:
